object width,height 타입 변경(int -> float)

+++ src/main/java/com/example/Mas/model/MrObject.java
@@ -30,8 +30,8 @@ public class MrObject implements Serializable {
     private Float x;
     private Float y;
     private Integer angle;
-    private Integer width;
-    private Integer height;
+    private Float width;
+    private Float height;
     private Integer objectSeq; // 저장된 순서
     @Column(name = "create_date", columnDefinition = "TIMESTAMP(3)")
     private Timestamp createDate; // 오브젝트 생성시 timestamp
